Last Commit
+Ability to change voxel Size
+grass rendering (off by default you will need to uncomment code)
+Wire frame mesh material
+New planet material based of UV's (solid method but its not done, wouldn't take long to add)
+Fixed XZ flipping on node spawning
+Terraforming base it works but is super slow and would be faster if someone made a new job that did the whole octree at once)
+Wind Map generation
+faster planet set-up (just add the script and fill the boxes then run it)
+ performance upgrades
+ Icons on hierarchy view (makes it look nice)
+Planet generation Reports
+Code reordering

not much removed but the "objectNoiseJob" would be easy to readd.

+++ Assets/Boxey/Planets/Core/Static/VoxelTables.cs
@@ -3,29 +3,30 @@ using Unity.Mathematics;
 namespace Boxey.Planets.Core.Static {
     public static class VoxelTables {
         public static readonly float3[] NodeOffsets = {
-            new (1f, 1f, 1f),
-            new (0f, 1f, 0f),
-            new (1f, 0f, 1f),
-            new (0f, 0f, 0f),
-            new (0f, 1f, 1f),
-            new (1f, 1f, 0f),
-            new (1f, 0f, 0f),
-            new (0f, 0f, 1f)
+            new(1f, 1f, 1f),
+            new(0f, 1f, 0f),
+            new(1f, 0f, 1f),
+            new(0f, 0f, 0f),
+            new(0f, 1f, 1f),
+            new(1f, 1f, 0f),
+            new(1f, 0f, 0f),
+            new(0f, 0f, 1f)
         };
-        public static readonly int3[] CornerTable = {
-            new (0, 0, 0),
-            new (1, 0, 0),
-            new (1, 1, 0),
-            new (0, 1, 0),
-            new (0, 0, 1),
-            new (1, 0, 1),
-            new (1, 1, 1),
-            new (0, 1, 1)
+
+        public static readonly float3[] CornerTable = {
+            new(0, 0, 0),
+            new(1, 0, 0),
+            new(1, 1, 0),
+            new(0, 1, 0),
+            new(0, 0, 1),
+            new(1, 0, 1),
+            new(1, 1, 1),
+            new(0, 1, 1)
         };
         public static readonly int[] EdgeIndexes = {
             0, 1, 1, 2, 3, 2, 0, 3, 4, 5, 5, 6, 7, 6, 4, 7, 0, 4, 1, 5, 2, 6, 3, 7
         };
-        public static readonly int[] TriangleTable = { 
+        public static readonly int[] TriangleTable = {
             -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             0, 8, 3, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             0, 1, 9, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
@@ -281,7 +282,7 @@ namespace Boxey.Planets.Core.Static {
             1, 3, 8, 9, 1, 8, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             0, 9, 1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
             0, 3, 8, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1 
+            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1
         };
     }
 }
